SEAccessKit_AR8.0.6_OpnSrc

+++ packages/EseSpi/jni/SpiChannel.h
@@ -23,7 +23,7 @@ typedef unsigned char   UINT8;
 typedef unsigned short  UINT16;
 typedef unsigned long   UINT32;
 typedef unsigned long long int UINT64;
-typedef signed   long   INT32;
+typedef int   INT32;
 typedef signed   char   INT8;
 typedef signed   short  INT16;
 typedef unsigned char   BOOLEAN;
